make sidesheets bottom bar sticky, cleanup and improvements
