+++ app/lib/features/events/widgets/create_event_sheet.dart
@@ -1,35 +1,38 @@
-import 'package:acter/common/themes/app_theme.dart';
+import 'package:acter/common/providers/space_providers.dart';
+import 'package:acter/common/snackbars/custom_msg.dart';
 import 'package:acter/common/utils/routes.dart';
 import 'package:acter/common/widgets/default_dialog.dart';
 import 'package:acter/common/widgets/input_text_field.dart';
 import 'package:acter/common/widgets/side_sheet.dart';
-import 'package:acter/features/events/providers/events_provider.dart';
-import 'package:acter_flutter_sdk/acter_flutter_sdk.dart';
+import 'package:acter/common/widgets/spaces/select_space_form_field.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import 'package:intl/intl.dart';
 
 // interface data providers
-final _titleProvider = StateProvider<String>((ref) => '');
-final _dateProvider = StateProvider<DateTime>((ref) => DateTime.now());
-final _startTimeProvider = StateProvider<TimeOfDay>((ref) => TimeOfDay.now());
-final _endTimeProvider = StateProvider(
+final _titleProvider = StateProvider.autoDispose<String>((ref) => '');
+final _dateProvider =
+    StateProvider.autoDispose<DateTime>((ref) => DateTime.now());
+final _startTimeProvider =
+    StateProvider.autoDispose<TimeOfDay>((ref) => TimeOfDay.now());
+final _endTimeProvider = StateProvider.autoDispose<TimeOfDay>(
   (ref) => TimeOfDay(
     hour: TimeOfDay.now().hour + 1,
     minute: TimeOfDay.now().minute,
   ),
 );
 
-class EditEventSheet extends ConsumerStatefulWidget {
-  final String? calendarId;
-  const EditEventSheet({super.key, this.calendarId});
+class CreateEventSheet extends ConsumerStatefulWidget {
+  final String? initialSelectedSpace;
+  const CreateEventSheet({super.key, this.initialSelectedSpace});
 
   @override
-  ConsumerState<EditEventSheet> createState() => _EditEventSheetConsumerState();
+  ConsumerState<CreateEventSheet> createState() =>
+      _CreateEventSheetConsumerState();
 }
 
-class _EditEventSheetConsumerState extends ConsumerState<EditEventSheet> {
+class _CreateEventSheetConsumerState extends ConsumerState<CreateEventSheet> {
   final TextEditingController _nameController = TextEditingController();
   final TextEditingController _descriptionController = TextEditingController();
   final TextEditingController _dateController = TextEditingController();
@@ -40,59 +43,23 @@ class _EditEventSheetConsumerState extends ConsumerState<EditEventSheet> {
   @override
   void initState() {
     super.initState();
-    _editEventData();
-  }
-
-  // apply existing data to fields
-  void _editEventData() async {
-    final calendarEvent = await ref.read(
-      calendarEventProvider(widget.calendarId!).future,
-    );
-    final titleNotifier = ref.read(_titleProvider.notifier);
-    final dateNotifier = ref.read(_dateProvider.notifier);
-    final startTimeNotifier = ref.read(_startTimeProvider.notifier);
-    final endTimeNotifier = ref.read(_endTimeProvider.notifier);
-
-    titleNotifier.update((state) => calendarEvent.title());
-    // parse RFC3393 date time
-    final dartDateTime = toDartDatetime(calendarEvent.utcStart());
-    final dartEndTime = toDartDatetime(calendarEvent.utcEnd());
-    dateNotifier.update(
-      (state) => DateTime(
-        dartDateTime.year,
-        dartDateTime.month,
-        dartDateTime.day,
-      ),
-    );
-    startTimeNotifier.update((state) => TimeOfDay.fromDateTime(dartDateTime));
-    endTimeNotifier.update((state) => TimeOfDay.fromDateTime(dartEndTime));
-
-    _nameController.text = ref.read(_titleProvider);
-    _dateController.text = DateFormat.yMd().format(ref.read(_dateProvider));
-
-    // We are doing as expected, but the lints triggers.
-    // ignore: use_build_context_synchronously
-    if (!context.mounted) {
-      return;
-    }
-    _startTimeController.text = ref.read(_startTimeProvider).format(context);
-    _endTimeController.text = ref.read(_endTimeProvider).format(context);
-    _descriptionController.text = calendarEvent.description()!.body();
+    WidgetsBinding.instance.addPostFrameCallback((Duration duration) {
+      final parentNotifier = ref.read(selectedSpaceIdProvider.notifier);
+      parentNotifier.state = widget.initialSelectedSpace;
+    });
   }
 
   @override
   Widget build(BuildContext context) {
     final titleInput = ref.watch(_titleProvider);
     return SideSheet(
-      header: 'Edit event',
+      header: 'Create new event',
       addActions: true,
       body: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 15, vertical: 10),
+        padding: const EdgeInsets.symmetric(horizontal: 15),
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: <Widget>[
-            const Text('Here you can edit event information'),
-            const SizedBox(height: 15),
             Column(
               crossAxisAlignment: CrossAxisAlignment.start,
               children: <Widget>[
@@ -243,50 +210,22 @@ class _EditEventSheetConsumerState extends ConsumerState<EditEventSheet> {
                   textInputType: TextInputType.url,
                   maxLines: 1,
                 ),
+                const SizedBox(height: 15),
+                const SelectSpaceFormField(canCheck: 'CanPostEvent'),
               ],
             ),
           ],
         ),
       ),
-      actions: [
-        ElevatedButton(
-          onPressed: () => context.canPop()
-              ? context.pop()
-              : context.goNamed(Routes.main.name),
-          style: ElevatedButton.styleFrom(
-            backgroundColor: Theme.of(context).colorScheme.neutral,
-            shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.circular(6),
-            ),
-            side: BorderSide(
-              color: Theme.of(context).colorScheme.success,
-            ),
-            foregroundColor: Theme.of(context).colorScheme.neutral6,
-            textStyle: Theme.of(context).textTheme.bodySmall,
-          ),
-          child: const Text('Cancel'),
-        ),
-        const SizedBox(width: 10),
-        ElevatedButton(
-          onPressed: () async {
-            if (titleInput.isEmpty) {
-              return;
-            }
-            _handleUpdateEvent(context);
-          },
-          style: ElevatedButton.styleFrom(
-            backgroundColor: titleInput.isNotEmpty
-                ? Theme.of(context).colorScheme.success
-                : Theme.of(context).colorScheme.success.withOpacity(0.6),
-            shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.circular(6),
-            ),
-            foregroundColor: Theme.of(context).colorScheme.neutral6,
-            textStyle: Theme.of(context).textTheme.bodySmall,
-          ),
-          child: const Text('Save Changes'),
-        ),
-      ],
+      confirmActionTitle: 'Create Event',
+      cancelActionTitle: 'Cancel',
+      confirmActionOnPressed: titleInput.isEmpty
+          ? null
+          : () async {
+              await _handleCreateEvent();
+            },
+      cancelActionOnPressed: () =>
+          context.canPop() ? context.pop() : context.goNamed(Routes.main.name),
     );
   }
 
@@ -294,26 +233,36 @@ class _EditEventSheetConsumerState extends ConsumerState<EditEventSheet> {
     ref.read(_titleProvider.notifier).update((state) => value!);
   }
 
-  void _handleUpdateEvent(BuildContext context) async {
+  Future<void> _handleCreateEvent() async {
     showAdaptiveDialog(
       barrierDismissible: false,
       context: context,
-      builder: (context) => DefaultDialog(
-        title: Text(
-          'Updating Event',
-          style: Theme.of(context).textTheme.titleSmall,
-        ),
+      builder: (context) => const DefaultDialog(
+        title: Text('Creating Event'),
         isLoader: true,
       ),
     );
-    final calendarEvent =
-        await ref.read(calendarEventProvider(widget.calendarId!).future);
+    // pre fill values if user doesn't set date time.
+    if (_dateController.text.isEmpty) {
+      _dateController.text = DateFormat.yMd().format(ref.read(_dateProvider));
+    }
+    if (_startTimeController.text.isEmpty) {
+      final time = ref.read(_startTimeProvider).format(context);
+      _startTimeController.text = time;
+    }
+    if (_endTimeController.text.isEmpty) {
+      final time = ref.read(_endTimeProvider).format(context);
+      _endTimeController.text = time;
+    }
     try {
-      // initialize event update builder
-      final eventUpdateBuilder = calendarEvent.updateBuilder();
+      final spaceId = ref.read(selectedSpaceIdProvider);
+      final space = await ref.read(spaceProvider(spaceId!).future);
+      final draft = space.calendarEventDraft();
 
-      eventUpdateBuilder.title(ref.read(_titleProvider));
+      draft.title(ref.read(_titleProvider));
+      draft.descriptionText(_descriptionController.text.trim());
 
+      // convert selected date time to utc and RFC3339 format
       final date = ref.read(_dateProvider);
       final startTime = ref.read(_startTimeProvider);
       final utcStartDateTime = DateTime(
@@ -323,8 +272,7 @@ class _EditEventSheetConsumerState extends ConsumerState<EditEventSheet> {
         startTime.hour,
         startTime.minute,
       ).toUtc();
-      eventUpdateBuilder
-          .utcStartFromRfc3339(utcStartDateTime.toIso8601String());
+      draft.utcStartFromRfc3339(utcStartDateTime.toIso8601String());
 
       final endTime = ref.read(_endTimeProvider);
       final utcEndDateTime = DateTime(
@@ -334,13 +282,10 @@ class _EditEventSheetConsumerState extends ConsumerState<EditEventSheet> {
         endTime.hour,
         endTime.minute,
       ).toUtc();
-      eventUpdateBuilder.utcEndFromRfc3339(utcEndDateTime.toIso8601String());
+      draft.utcEndFromRfc3339(utcEndDateTime.toIso8601String());
 
-      eventUpdateBuilder.descriptionText(_descriptionController.text.trim());
-
-      final eventId = await eventUpdateBuilder.send();
-      debugPrint('Updated Calendar Event: ${eventId.toString()}');
-
+      final eventId = await draft.send();
+      debugPrint('Created Calendar Event: ${eventId.toString()}');
       // We are doing as expected, but the lints triggers.
       // ignore: use_build_context_synchronously
       if (!context.mounted) {
@@ -348,14 +293,16 @@ class _EditEventSheetConsumerState extends ConsumerState<EditEventSheet> {
       }
       context.pop();
       context.pop();
+      await context.pushNamed(
+        Routes.calendarEvent.name,
+        pathParameters: {'calendarId': eventId.toString()},
+      );
     } catch (e) {
-      // We are doing as expected, but the lints triggers.
-      // ignore: use_build_context_synchronously
       if (!context.mounted) {
         return;
       }
       context.pop();
-      debugPrint('Some error occured ${e.toString()}');
+      customMsgSnackbar(context, 'Some error occurred $e');
     }
   }
 
@@ -378,7 +325,13 @@ class _EditEventSheetConsumerState extends ConsumerState<EditEventSheet> {
       context: context,
       initialTime: ref.read(_startTimeProvider),
     );
-    if (picked != null && context.mounted) {
+
+    // We are doing as expected, but the lints triggers.
+    // ignore: use_build_context_synchronously
+    if (!context.mounted) {
+      return;
+    }
+    if (picked != null) {
       ref.read(_startTimeProvider.notifier).update((state) => picked);
       final time = ref.read(_startTimeProvider).format(context);
       _startTimeController.text = time;
@@ -390,7 +343,13 @@ class _EditEventSheetConsumerState extends ConsumerState<EditEventSheet> {
       context: context,
       initialTime: ref.read(_endTimeProvider),
     );
-    if (picked != null && context.mounted) {
+
+    // We are doing as expected, but the lints triggers.
+    // ignore: use_build_context_synchronously
+    if (!context.mounted) {
+      return;
+    }
+    if (picked != null) {
       ref.read(_endTimeProvider.notifier).update((state) => picked);
       final time = ref.read(_endTimeProvider).format(context);
       _endTimeController.text = time;
