--- app/lib/features/pins/dialogs/create_pin_sheet.dart
@@ -1,187 +1,0 @@
-import 'package:acter/common/providers/space_providers.dart';
-import 'package:acter/common/snackbars/custom_msg.dart';
-import 'package:acter/common/themes/app_theme.dart';
-import 'package:acter/common/utils/routes.dart';
-import 'package:acter/common/widgets/default_dialog.dart';
-import 'package:acter/common/widgets/md_editor_with_preview.dart';
-import 'package:acter/common/widgets/side_sheet.dart';
-import 'package:acter/common/widgets/spaces/select_space_form_field.dart';
-import 'package:atlas_icons/atlas_icons.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:go_router/go_router.dart';
-
-// interface data providers
-final titleProvider = StateProvider<String>((ref) => '');
-final textProvider = StateProvider<String>((ref) => '');
-final linkProvider = StateProvider<String>((ref) => '');
-
-class CreatePinSheet extends ConsumerStatefulWidget {
-  final String? initialSelectedSpace;
-  const CreatePinSheet({super.key, this.initialSelectedSpace});
-
-  @override
-  ConsumerState<CreatePinSheet> createState() => _CreatePinSheetConsumerState();
-}
-
-class _CreatePinSheetConsumerState extends ConsumerState<CreatePinSheet> {
-  final TextEditingController _titleController = TextEditingController();
-  final TextEditingController _linkController = TextEditingController();
-  final TextEditingController _textController = TextEditingController();
-  final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
-
-  @override
-  void initState() {
-    super.initState();
-    WidgetsBinding.instance.addPostFrameCallback((Duration duration) {
-      final spaceNotifier = ref.read(selectedSpaceIdProvider.notifier);
-      spaceNotifier.state = widget.initialSelectedSpace;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    bool hasLinkOrText() {
-      return _linkController.text.isNotEmpty || _textController.text.isNotEmpty;
-    }
-
-    return SideSheet(
-      header: 'Create new Pin',
-      addActions: true,
-      body: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 15, vertical: 10),
-        child: Form(
-          key: _formKey,
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: <Widget>[
-              Padding(
-                padding: const EdgeInsets.symmetric(vertical: 10),
-                child: Row(
-                  children: [
-                    Expanded(
-                      child: SizedBox(
-                        height: 52,
-                        child: TextFormField(
-                          decoration: InputDecoration(
-                            hintText: 'Your title',
-                            labelText: 'Title',
-                            border: OutlineInputBorder(
-                              borderRadius: BorderRadius.circular(5),
-                            ),
-                          ),
-                          controller: _titleController,
-                          validator: (value) =>
-                              (value != null && value.isNotEmpty)
-                                  ? null
-                                  : 'Please enter a title',
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-              TextFormField(
-                decoration: InputDecoration(
-                  icon: const Icon(Atlas.link_thin),
-                  hintText: 'https://',
-                  labelText: 'link',
-                  border: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(5),
-                  ),
-                ),
-                controller: _linkController,
-                validator: (value) =>
-                    hasLinkOrText() ? null : 'Text or URL must be given',
-              ),
-              MdEditorWithPreview(
-                validator: (value) =>
-                    hasLinkOrText() ? null : 'Text or URL must be given',
-                controller: _textController,
-              ),
-              const SelectSpaceFormField(canCheck: 'CanPostPin'),
-            ],
-          ),
-        ),
-      ),
-      actions: [
-        ElevatedButton(
-          onPressed: () => context.canPop()
-              ? context.pop()
-              : context.goNamed(Routes.main.name),
-          style: ElevatedButton.styleFrom(
-            shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.circular(6),
-            ),
-            foregroundColor: Theme.of(context).colorScheme.neutral6,
-            textStyle: Theme.of(context).textTheme.bodySmall,
-          ),
-          child: const Text('Cancel'),
-        ),
-        const SizedBox(width: 10),
-        ElevatedButton(
-          onPressed: () async {
-            if (_formKey.currentState!.validate()) {
-              showAdaptiveDialog(
-                barrierDismissible: false,
-                context: context,
-                builder: (context) => DefaultDialog(
-                  title: Text(
-                    'Posting Pin',
-                    style: Theme.of(context).textTheme.titleSmall,
-                  ),
-                  isLoader: true,
-                ),
-              );
-              try {
-                final spaceId = ref.read(selectedSpaceIdProvider);
-                final space = await ref.read(spaceProvider(spaceId!).future);
-                final pinDraft = space.pinDraft();
-                final text = _textController.text;
-                final url = _linkController.text;
-                pinDraft.title(ref.read(titleProvider));
-                if (text.isNotEmpty) {
-                  pinDraft.contentMarkdown(text);
-                }
-                if (url.isNotEmpty) {
-                  pinDraft.url(url);
-                }
-                final pinId = await pinDraft.send();
-                // reset providers
-                _textController.text = '';
-                _linkController.text = '';
-
-                // We are doing as expected, but the lints triggers.
-                // ignore: use_build_context_synchronously
-                if (!context.mounted) {
-                  return;
-                }
-                Navigator.of(context, rootNavigator: true).pop();
-                context.pushNamed(
-                  Routes.pin.name,
-                  pathParameters: {'pinId': pinId.toString()},
-                );
-              } catch (e) {
-                // We are doing as expected, but the lints triggers.
-                // ignore: use_build_context_synchronously
-                if (!context.mounted) {
-                  return;
-                }
-                customMsgSnackbar(context, 'Failed to pin: $e');
-              }
-            }
-          },
-          style: ElevatedButton.styleFrom(
-            backgroundColor: Theme.of(context).colorScheme.success,
-            shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.circular(6),
-            ),
-            foregroundColor: Theme.of(context).colorScheme.neutral6,
-            textStyle: Theme.of(context).textTheme.bodySmall,
-          ),
-          child: const Text('Create Pin'),
-        ),
-      ],
-    );
-  }
-}
