+++ app/lib/router/router.dart
@@ -9,17 +9,17 @@ import 'package:acter/features/events/pages/events_page.dart';
 import 'package:acter/features/settings/pages/blocked_users.dart';
 import 'package:acter/features/settings/pages/sessions_page.dart';
 import 'package:acter/features/bug_report/pages/bug_report_page.dart';
-import 'package:acter/features/chat/dialogs/create_chat_sheet.dart';
+import 'package:acter/features/chat/widgets/create_chat_sheet.dart';
 import 'package:acter/features/chat/pages/chat_select_page.dart';
 import 'package:acter/features/chat/pages/chats_shell.dart';
 import 'package:acter/features/chat/pages/room_profile_page.dart';
 import 'package:acter/features/chat/pages/room_page.dart';
-import 'package:acter/features/events/dialogs/create_event_sheet.dart';
-import 'package:acter/features/events/dialogs/edit_event_sheet.dart';
+import 'package:acter/features/events/widgets/create_event_sheet.dart';
+import 'package:acter/features/events/widgets/edit_event_sheet.dart';
 import 'package:acter/features/events/pages/event_page.dart';
 import 'package:acter/features/home/pages/dashboard.dart';
 import 'package:acter/features/home/pages/home_shell.dart';
-import 'package:acter/features/pins/dialogs/edit_pin_sheet.dart';
+import 'package:acter/features/pins/widgets/edit_pin_sheet.dart';
 import 'package:acter/features/space/pages/members_page.dart';
 import 'package:acter/features/news/pages/news_page.dart';
 import 'package:acter/features/news/pages/simple_post.dart';
@@ -28,7 +28,7 @@ import 'package:acter/features/onboarding/pages/intro_profile.dart';
 import 'package:acter/features/onboarding/pages/login_page.dart';
 import 'package:acter/features/onboarding/pages/register_page.dart';
 import 'package:acter/features/onboarding/pages/start_page.dart';
-import 'package:acter/features/pins/dialogs/create_pin_sheet.dart';
+import 'package:acter/features/pins/widgets/create_pin_sheet.dart';
 import 'package:acter/features/pins/pages/pin_page.dart';
 import 'package:acter/features/pins/pages/pins_page.dart';
 import 'package:acter/features/profile/pages/my_profile_page.dart';
@@ -38,7 +38,7 @@ import 'package:acter/features/settings/pages/index_page.dart';
 import 'package:acter/features/settings/pages/info_page.dart';
 import 'package:acter/features/settings/pages/labs_page.dart';
 import 'package:acter/features/settings/pages/licenses_page.dart';
-import 'package:acter/features/space/dialogs/edit_space_sheet.dart';
+import 'package:acter/features/space/widgets/edit_space_sheet.dart';
 import 'package:acter/common/dialogs/invite_to_room_dialog.dart';
 import 'package:acter/features/space/pages/chats_page.dart';
 import 'package:acter/features/space/pages/events_page.dart';
@@ -52,7 +52,7 @@ import 'package:acter/features/tasks/dialogs/create_task_list_sheet.dart';
 import 'package:acter/features/tasks/pages/tasks_page.dart';
 import 'package:acter/features/space/settings/pages/apps_settings_page.dart';
 import 'package:acter/features/space/settings/pages/index_page.dart';
-import 'package:acter/features/spaces/dialogs/create_space_sheet.dart';
+import 'package:acter/features/spaces/widgets/create_space_sheet.dart';
 import 'package:acter/features/spaces/pages/join_space.dart';
 import 'package:acter/features/spaces/pages/spaces_page.dart';
 import 'package:acter_flutter_sdk/acter_flutter_sdk.dart';
